net/smc: embedded tcp sock into smc sock
For inet version of SMC, one of the key goals is to make
a fallbacked smc sock can be recognazied as a tcp sock by
net tools.

So, it is necessary to embedded the tcp sock into smc sock
and make the tcp sock as the first member of smc sock.

Signed-off-by: D. Wythe <alibuda@linux.alibaba.com>