net/smc: support smc inet with merge socks
This patch mainly implements how AF_INET works under one single
sock.

Unlike the AF_SMC sock, the inet smc can directly fallback to TCP
under the IRQ context, no need to start a work in workqueue. The
inet smc start the listen work only when the peer has syn_smc set.
It accelerates the efficiency of establishing connections whom prove
to be SMC unsupported.

The complex logic here is that we need to dynamically sort the reqsk in
TCP, and try to move ahead the syn_smc with 0 as much as possible, which
allows for faster fallback connections. However, due to the timing issue
of accept, we cannot always guarantee this condition.

Signed-off-by: D. Wythe <alibuda@linux.alibaba.com>